+++ types/v56/types.go
@@ -985,8 +985,8 @@ type OrgLdapGroupAttributes struct {
 	ObjectIdentifier     string `xml:"ObjectIdentifier"`
 	GroupName            string `xml:"GroupName"`
 	Membership           string `xml:"Membership"`
-	BackLinkIdentifier   string `xml:"BackLinkIdentifier,omitempty"`
 	MembershipIdentifier string `xml:"MembershipIdentifier"`
+	BackLinkIdentifier   string `xml:"BackLinkIdentifier,omitempty"`
 }
 
 // OrgLdapUserAttributesType represents the ldap user attribute settings for a VMware Cloud Director organization.
@@ -3142,6 +3142,7 @@ type CatalogRecord struct {
 	Description             string    `xml:"description,attr,omitempty"`
 	IsPublished             bool      `xml:"isPublished,attr,omitempty"`
 	IsShared                bool      `xml:"isShared,attr,omitempty"`
+	IsLocal                 bool      `xml:"isLocal,attr,omitempty"`
 	CreationDate            string    `xml:"creationDate,attr,omitempty"`
 	OrgName                 string    `xml:"orgName,attr,omitempty"`
 	OwnerName               string    `xml:"ownerName,attr,omitempty"`